refactor(stdlib): split workflow.rs into workflow/ module
Breaks the 2240-line crates/harn-vm/src/stdlib/workflow.rs into focused
submodules: convert, policy, usage, stage, map, artifact, guards,
register, plus tests. Largest resulting file is register.rs at 923
lines (was 2240). Behavior is unchanged; workflow builtins,
execute_workflow, load_run_tree, and all 14 module-level tests retain
their call paths via mod.rs re-exports.

Closes #45.
