+++ crates/harn-vm/src/stdlib/workflow/mod.rs
@@ -1,0 +1,17 @@
+//! Workflow graph manipulation and execution builtins.
+
+mod artifact;
+mod convert;
+mod guards;
+mod map;
+mod policy;
+mod register;
+mod stage;
+mod usage;
+
+pub(in crate::stdlib) use self::artifact::load_run_tree;
+pub(in crate::stdlib) use self::register::execute_workflow;
+pub(crate) use self::register::register_workflow_builtins;
+
+#[cfg(test)]
+mod tests;
