+++ crates/harn-vm/src/stdlib/workflow/tests.rs
@@ -1,4 +1,6 @@
-use super::*;
+use super::artifact::{load_run_tree, snapshot_trace_spans};
+use super::map::{execute_join_policy, LocalTask};
+use super::stage::{classify_stage_outcome, execute_stage_attempts};
 use crate::orchestration::{render_artifacts_context, render_workflow_prompt};
 use crate::orchestration::{save_run_record, RunChildRecord, RunRecord};
 use crate::tracing::{set_tracing_enabled, span_end, span_start, SpanKind};
